[datatypes] Reorganize
